Worked on Vertical Slice Level
Added entrance door and glyph.
Layed out NavMeshVolumes.
Added Howler spawn trigger in tunnel.
Added 2 glyphs and corresponding events that spawn drones on each side.
Added gaps in floor.
Added 2nd door which needs 3 glyphs, but the order of input doesn't correspond to the order of the numbers.
Working on that for next push.